+++ CHANGES.rst
@@ -5,14 +5,16 @@
 - fix get_rank_genes_tables return less than n_genes for filtered ranking
 - fix _search_dim_red_parameters "ValueError: 'left' is not a valid value for loc"
 - fix gseapy, louvain install outside of the docker image (#310)
+- implemented plotting.embedding.agg_feature_embedding
 - cleanup installation/ dependencies
 - CI: Revert to installing mampok from main
 - implement bgcolor cell-selection
-- add a bar plot to predict_cell_cycle (#301)
 - update readme
+- add a bar plot to predict_cell_cycle (#301)
 
 Changes to notebooks
 ^^^^^^^^^^^^^^^^^^^^
+- add agg_feature_embedding to the group_marker notebook
 - fix RNA:02 zebrafish gender suggestion (#311)
 - revise 03_normalization_batch_correction notebook
 - prepare_for_cellxgene: Add auth parameter